# opsplatform-video-manager-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CDNProvider:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 1
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  models.CreateTokenRequest:
    properties:
      expires_in:
        description: Expiration time in seconds (ignored if never_expire is true)
        type: integer
      name:
        description: Token name/description
        maxLength: 255
        minLength: 1
        type: string
      never_expire:
        description: If true, token never expires
        type: boolean
    required:
    - name
    type: object
...
  models.LoginRequest:
    properties:
      password:
        minLength: 1
        type: string
      username:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      is_admin:
        type: boolean
      token:
        type: string
      username:
        type: string
    type: object
...
  models.Token:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      name:
        type: string
      never_expire:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Video Manager API Documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Video Manager API
  version: "1.0"
paths:
  /api/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Password change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: User login
      tags:
      - auth
  /api/auth/me:
    get:
      description: Get current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/auth/oidc/callback:
    get:
      description: Handle OIDC callback, create/login local user as non-admin, then
        redirect to frontend login
      parameters:
      - description: OIDC authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OIDC state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to frontend with token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.Response'
      summary: OIDC login callback
      tags:
      - auth
  /api/auth/oidc/login:
    get:
      description: Redirect to OIDC provider for SSO login
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to OIDC provider
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.Response'
      summary: OIDC login redirect
      tags:
      - auth
  /api/auth/token-info:
    get:
      description: Get current JWT token information including expiration time
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get token information
      tags:
      - auth
  /api/auth/tokens:
    get:
      description: Get all tokens created by the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Token'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all tokens
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: Create a new JWT token with custom expiration
      parameters:
      - description: Token creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create token
      tags:
      - auth
  /api/auth/tokens/{id}:
    delete:
      description: Delete a token by ID (only if it belongs to the current user)
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete token
      tags:
      - auth
  /api/cdn-lines:
    get:
      description: Retrieve all CDN lines, optionally filtered by provider ID
      parameters:
      - description: Filter by provider ID
        in: query
        name: provider_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CDNLine'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all CDN lines
      tags:
      - cdn-lines
    post:
      consumes:
      - application/json
      description: Create a new CDN line with the provided information
      parameters:
      - description: CDN Line information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCDNLineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNLine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new CDN line
      tags:
      - cdn-lines
  /api/cdn-lines/{id}:
    delete:
      description: Delete a CDN line by its ID. This will also delete associated video
        stream endpoints.
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a CDN line
      tags:
      - cdn-lines
    get:
      description: Retrieve a CDN line by its ID
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNLine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get CDN line by ID
      tags:
      - cdn-lines
    put:
      consumes:
      - application/json
      description: Update an existing CDN line by its ID
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated CDN Line information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCDNLineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNLine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a CDN line
      tags:
      - cdn-lines
  /api/cdn-providers:
    get:
      description: Retrieve all CDN providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CDNProvider'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all CDN providers
      tags:
      - cdn-providers
    post:
      consumes:
      - application/json
      description: Create a new CDN provider with the provided information
      parameters:
      - description: CDN Provider information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCDNProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNProvider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new CDN provider
      tags:
      - cdn-providers
...
  /api/cdn-providers/{id}/lines:
    get:
      description: Retrieve all CDN lines associated with a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CDNLine'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get CDN lines by provider
      tags:
      - cdn-providers
...
  /api/stats:
    get:
      description: Retrieve comprehensive system statistics including counts of providers,
        lines, domains, streams, stream paths, endpoints, and various aggregations
      produces:
      - application/json
      responses:
        "200":
          description: Statistics object containing counts and aggregations
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get system statistics
      tags:
      - stats
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"